viewer/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
asyncpg
python-jose[cryptography]
bcrypt
httpx
pydantic[email]
anthropic
